move encoding tests from pii endpoint to separate file
