+++ sdk/textanalytics/azure-ai-textanalytics/tests/recordings/test_encoding.test_diacritics_nfd.yaml
@@ -1,7 +1,7 @@
 interactions:
 - request:
-    body: '{"documents": [{"id": "0", "text": "\ud83d\udc69\ud83c\udffb SSN: 859-98-0987",
-      "language": "en"}]}'
+    body: '{"documents": [{"id": "0", "text": "an\u0303o SSN: 859-98-0987", "language":
+      "en"}]}'
     headers:
       Accept:
       - application/json, text/json
@@ -10,7 +10,7 @@ interactions:
       Connection:
       - keep-alive
       Content-Length:
-      - '99'
+      - '84'
       Content-Type:
       - application/json
       User-Agent:
@@ -20,16 +20,16 @@ interactions:
   response:
     body:
       string: '{"documents":[{"id":"0","entities":[{"text":"859-98-0987","category":"U.S.
-        Social Security Number (SSN)","offset":8,"length":11,"confidenceScore":0.65}],"warnings":[]}],"errors":[],"modelVersion":"2020-07-01"}'
+        Social Security Number (SSN)","offset":10,"length":11,"confidenceScore":0.65}],"warnings":[]}],"errors":[],"modelVersion":"2020-07-01"}'
     headers:
       apim-request-id:
-      - 12d4e642-e7ca-4549-83f3-37bd93997b0d
+      - 097b0d6d-4a0b-42eb-894f-a40c40898515
       content-type:
       - application/json; charset=utf-8
       csp-billing-usage:
       - CognitiveServices.TextAnalytics.BatchScoring=1
       date:
-      - Thu, 27 Aug 2020 19:32:19 GMT
+      - Fri, 28 Aug 2020 16:54:37 GMT
       strict-transport-security:
       - max-age=31536000; includeSubDomains; preload
       transfer-encoding:
@@ -37,7 +37,7 @@ interactions:
       x-content-type-options:
       - nosniff
       x-envoy-upstream-service-time:
-      - '98'
+      - '65'
     status:
       code: 200
       message: OK
