+++ sdk/textanalytics/azure-ai-textanalytics/tests/recordings/test_encoding.test_korean_nfc.yaml
@@ -1,6 +1,6 @@
 interactions:
 - request:
-    body: '{"documents": [{"id": "0", "text": "a\u00f1o SSN: 859-98-0987", "language":
+    body: '{"documents": [{"id": "0", "text": "\uc544\uac00 SSN: 859-98-0987", "language":
       "en"}]}'
     headers:
       Accept:
@@ -10,7 +10,7 @@ interactions:
       Connection:
       - keep-alive
       Content-Length:
-      - '83'
+      - '87'
       Content-Type:
       - application/json
       User-Agent:
@@ -20,16 +20,16 @@ interactions:
   response:
     body:
       string: '{"documents":[{"id":"0","entities":[{"text":"859-98-0987","category":"U.S.
-        Social Security Number (SSN)","offset":9,"length":11,"confidenceScore":0.65}],"warnings":[]}],"errors":[],"modelVersion":"2020-07-01"}'
+        Social Security Number (SSN)","offset":8,"length":11,"confidenceScore":0.65}],"warnings":[]}],"errors":[],"modelVersion":"2020-07-01"}'
     headers:
       apim-request-id:
-      - e3989986-663b-4ca0-a936-05cad3ab7e79
+      - 8a681705-efe1-45e4-a9d5-d94ed8b4180d
       content-type:
       - application/json; charset=utf-8
       csp-billing-usage:
       - CognitiveServices.TextAnalytics.BatchScoring=1
       date:
-      - Thu, 27 Aug 2020 19:32:17 GMT
+      - Fri, 28 Aug 2020 16:54:39 GMT
       strict-transport-security:
       - max-age=31536000; includeSubDomains; preload
       transfer-encoding:
@@ -37,7 +37,7 @@ interactions:
       x-content-type-options:
       - nosniff
       x-envoy-upstream-service-time:
-      - '82'
+      - '90'
     status:
       code: 200
       message: OK
