+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_recognize_pii_entities.py
@@ -87,71 +87,6 @@ class TestRecognizePIIEntities(TextAnalyticsTest):
 
     @GlobalTextAnalyticsAccountPreparer()
     @TextAnalyticsClientPreparer()
-    def test_emoji(self, client):
-        result = client.recognize_pii_entities(["👩 SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 7)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_emoji_with_skin_tone_modifier(self, client):
-        result = client.recognize_pii_entities(["👩🏻 SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 8)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_emoji_family(self, client):
-        result = client.recognize_pii_entities(["👩‍👩‍👧‍👧 SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 13)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_emoji_family_with_skin_tone_modifier(self, client):
-        result = client.recognize_pii_entities(["👩🏻‍👩🏽‍👧🏾‍👦🏿 SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 17)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_diacritics_nfc(self, client):
-        result = client.recognize_pii_entities(["año SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 9)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_diacritics_nfd(self, client):
-        result = client.recognize_pii_entities(["año SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 10)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_korean_nfc(self, client):
-        result = client.recognize_pii_entities(["아가 SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 8)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_korean_nfd(self, client):
-        result = client.recognize_pii_entities(["아가 SSN: 859-98-0987"])
-        self.assertEqual(result[0].entities[0].offset, 8)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
-    def test_zalgo_text(self, client):
-        result = client.recognize_pii_entities(["ơ̵̧̧̢̳̘̘͕͔͕̭̟̙͎͈̞͔̈̇̒̃͋̇̅͛̋͛̎́͑̄̐̂̎͗͝m̵͍͉̗̄̏͌̂̑̽̕͝͠g̵̢̡̢̡̨̡̧̛͉̞̯̠̤̣͕̟̫̫̼̰͓̦͖̣̣͎̋͒̈́̓̒̈̍̌̓̅͑̒̓̅̅͒̿̏́͗̀̇͛̏̀̈́̀̊̾̀̔͜͠͝ͅ SSN: 859-98-0987"])
-
-
-        self.assertEqual(result[0].entities[0].offset, 121)
-        self.assertEqual(result[0].entities[0].length, 11)
-
-    @GlobalTextAnalyticsAccountPreparer()
-    @TextAnalyticsClientPreparer()
     def test_passing_only_string(self, client):
         docs = [
             u"My SSN is 859-98-0987.",
